Refactored test and cleaned up initial example
