+++ src/AzurePipelinesToGitHubActionsConverter.Tests/StepsScriptingTests.cs
@@ -1,0 +1,350 @@
+﻿using AzurePipelinesToGitHubActionsConverter.Core;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
+
+namespace AzurePipelinesToGitHubActionsConverter.Tests
+{
+    [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
+    [TestClass]
+    public class StepsScriptingTests
+    {
+        [TestMethod]
+        public void BashIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- bash: echo 'some text'
+  displayName: test bash
+";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: test bash
+  run: echo 'some text'
+  shell: bash
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+        [TestMethod]
+        public void BashTaskIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- task: Bash@3
+  displayName: test bash
+  inputs:
+    targetType: 'inline'
+    script: echo $MYSECRET
+  env:
+    MYSECRET: $(Foo)
+";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: test bash
+  run: echo $MYSECRET
+  shell: bash
+  env:
+    MYSECRET: ${{ env.Foo }}
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+        [TestMethod]
+        public void BashWithBatchShellTaskIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- task: ShellScript@2
+  displayName: test bash
+  inputs:
+    scriptPath: myscript.sh
+    args: -f 'John Smith'
+";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: test bash
+  run: myscript.sh -f 'John Smith'
+  shell: bash
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+              [TestMethod]
+        public void BatchScriptIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+  - task: BatchScript@1
+    inputs:
+      filename: scripts/azure-pipelines/setup_msvc_env.bat
+      modifyEnvironment: True
+      arguments: /w
+    condition: eq(variables['Agent.OS'], 'Windows_NT')
+    displayName: Setup MSVC Environment";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: Setup MSVC Environment
+  run: scripts/azure-pipelines/setup_msvc_env.bat /w
+  shell: cmd
+  if: (runner.os == 'Windows_NT')
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+        [TestMethod]
+        public void CmdLineScriptIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- task: CmdLine@2
+  inputs:
+    script: echo your commands here 
+";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- run: echo your commands here
+  shell: cmd
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+        [TestMethod]
+        public void MultilineScriptIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- script: |
+    echo Add other tasks to build, test, and deploy your project.
+    echo See https://aka.ms/yaml
+  displayName: 'Run a multi-line script'";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: Run a multi-line script
+  run: |
+    echo Add other tasks to build, test, and deploy your project.
+    echo See https://aka.ms/yaml
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+        [TestMethod]
+        public void PowershellWithSinglelineIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- task: PowerShell@2
+  displayName: 'PowerShell test task'
+  inputs:
+    targetType: inline
+    script: Write-Host 'Hello World'";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: PowerShell test task
+  run: Write-Host 'Hello World'
+  shell: powershell
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+        [TestMethod]
+        public void PowershellWithMultilineIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- task: PowerShell@2
+  displayName: 'PowerShell test task'
+  inputs:
+    targetType: inline
+    script: |
+      Write-Host 'Hello World'
+      Write-Host 'Hello World2'";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: PowerShell test task
+  run: |
+    Write-Host 'Hello World'
+    Write-Host 'Hello World2'
+  shell: powershell
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+        [TestMethod]
+        public void PowershellWithFileIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+    - task: PowerShell@2
+      displayName: 'PowerShell test task'
+      inputs:
+        targetType: FilePath
+        filePath: MyProject/BuildVersion.ps1
+        arguments: -ProjectFile ""MyProject/MyProject.Web/MyProject.Web.csproj""
+";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: PowerShell test task
+  run: MyProject/BuildVersion.ps1 -ProjectFile ""MyProject/MyProject.Web/MyProject.Web.csproj""
+  shell: powershell
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+        [TestMethod]
+        public void PowershellWithConditionIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- task: PowerShell@2
+  displayName: 'PowerShell test task'
+  condition: and(eq('ABCDE', 'BCD'), ne(0, 1))
+  inputs:
+    targetType: inline
+    script: Write-Host 'Hello World'";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: PowerShell test task
+  run: Write-Host 'Hello World'
+  shell: powershell
+  if: (('ABCDE' == 'BCD') && (0 != 1))
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+   
+        [TestMethod]
+        public void InnerPowershellIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- task: InlinePowershell@1                
+  displayName: 'old Powershell task'
+  inputs:
+    Script: |
+      Write-Host 'Hello World'
+      Write-Host 'Hello World2'
+";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: old Powershell task
+  run: |
+    Write-Host 'Hello World'
+    Write-Host 'Hello World2'
+  shell: powershell
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+        [TestMethod]
+        public void PowerShellIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- powershell: Write-Host 'some text'
+  displayName: test PowerShell
+";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: test PowerShell
+  run: Write-Host 'some text'
+  shell: powershell
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+        [TestMethod]
+        public void PwshIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- pwsh: Write-Host 'some text'
+  displayName: test pwsh
+";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: test pwsh
+  run: Write-Host 'some text'
+  shell: pwsh
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+    }
+}
