+++ src/AzurePipelinesToGitHubActionsConverter.Tests/StepsTests.cs
@@ -8,25 +8,55 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
     [TestClass]
     public class StepsTests
     {
-
         [TestMethod]
-        public void InvalidStepIndividualStepTest()
+        public void AntStepTest()
         {
             //Arrange
             Conversion conversion = new Conversion();
-            string yaml = "- task: invalid fake task";
+            string yaml = @"
+- task: Ant@1
+  inputs:
+    workingDirectory: ''
+    buildFile: 'build.xml'
+    javaHomeOption: 'JDKVersion'
+    jdkVersionOption: '1.8'
+    jdkArchitectureOption: 'x64'
+    publishJUnitResults: true
+    testResultsFiles: '**/TEST-*.xml'  
+";
 
             //Act
             ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
 
             //Assert
             string expected = @"
-- # ""Error: the step 'invalid fake task' does not have a conversion path yet""
-  run: |
-    echo ""Error: the step 'invalid fake task' does not have a conversion path yet""
-    #task: invalid fake task
+- run: ant -noinput -buildfile build.xml
 ";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
 
+        [TestMethod]
+        public void ArchiveFilesStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- task: ArchiveFiles@2
+  inputs:
+    rootFolderOrFile: '$(build.sourcesDirectory)'
+    includeRootFolder: false";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- # 'Note: This is a third party action and currently only supports Linux: https://github.com/marketplace/actions/create-zip-file'
+  uses: montudor/action-zip@v0.1.0
+  with:
+    args: zip -qq -r  ${{ github.workspace }}
+";
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
         }
@@ -54,154 +84,6 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
   with:
     build-dir: build
     build-options: -A x64 -DCMAKE_TOOLCHAIN_FILE=../../vcpkg/scripts/buildsystems/vcpkg.cmake -DSELENE_BUILD_ALL=ON -DSELENE_WARNINGS_AS_ERRORS=ON ..
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void CmdLineIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: CmdLine@2
-  inputs:
-    script: echo your commands here 
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- run: echo your commands here
-  shell: cmd
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void PowerShellIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- powershell: Write-Host 'some text'
-  displayName: test PowerShell
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: test PowerShell
-  run: Write-Host 'some text'
-  shell: powershell
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void PwshIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- pwsh: Write-Host 'some text'
-  displayName: test pwsh
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: test pwsh
-  run: Write-Host 'some text'
-  shell: pwsh
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void BashIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- bash: echo 'some text'
-  displayName: test bash
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: test bash
-  run: echo 'some text'
-  shell: bash
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void BashTaskIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: Bash@3
-  displayName: test bash
-  inputs:
-    targetType: 'inline'
-    script: echo $MYSECRET
-  env:
-    MYSECRET: $(Foo)
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: test bash
-  run: echo $MYSECRET
-  shell: bash
-  env:
-    MYSECRET: ${{ env.Foo }}
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void BashWithBatchShellTaskIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: ShellScript@2
-  displayName: test bash
-  inputs:
-    scriptPath: myscript.sh
-    args: -f 'John Smith'
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: test bash
-  run: myscript.sh -f 'John Smith'
-  shell: bash
 ";
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
@@ -323,532 +205,6 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
         }
 
-
-        [TestMethod]
-        public void UseDotNetIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: UseDotNet@2
-  displayName: 'Use .NET Core sdk'
-  inputs:
-    packageType: sdk
-    version: 2.2.203";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Use .NET Core sdk
-  uses: actions/setup-dotnet@v1
-  with:
-    dotnet-version: 2.2.203
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-
-        [TestMethod]
-        public void BatchIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-  - task: BatchScript@1
-    inputs:
-      filename: scripts/azure-pipelines/setup_msvc_env.bat
-      modifyEnvironment: True
-      arguments: /w
-    condition: eq(variables['Agent.OS'], 'Windows_NT')
-    displayName: Setup MSVC Environment";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Setup MSVC Environment
-  run: scripts/azure-pipelines/setup_msvc_env.bat /w
-  shell: cmd
-  if: (runner.os == 'Windows_NT')
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void BuildDotNetIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- script: dotnet build --configuration $(buildConfiguration) WebApplication1/WebApplication1.Service/WebApplication1.Service.csproj
-  displayName: dotnet build $(buildConfiguration) part 1";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: dotnet build ${{ env.buildConfiguration }} part 1
-  run: dotnet build --configuration ${{ env.buildConfiguration }} WebApplication1/WebApplication1.Service/WebApplication1.Service.csproj
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-
-        [TestMethod]
-        public void PowershellWithSinglelineIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: PowerShell@2
-  displayName: 'PowerShell test task'
-  inputs:
-    targetType: inline
-    script: Write-Host 'Hello World'";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: PowerShell test task
-  run: Write-Host 'Hello World'
-  shell: powershell
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void PowershellWithMultilineIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: PowerShell@2
-  displayName: 'PowerShell test task'
-  inputs:
-    targetType: inline
-    script: |
-      Write-Host 'Hello World'
-      Write-Host 'Hello World2'";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: PowerShell test task
-  run: |
-    Write-Host 'Hello World'
-    Write-Host 'Hello World2'
-  shell: powershell
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-
-        [TestMethod]
-        public void PowershellWithFileIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-    - task: PowerShell@2
-      displayName: 'PowerShell test task'
-      inputs:
-        targetType: FilePath
-        filePath: MyProject/BuildVersion.ps1
-        arguments: -ProjectFile ""MyProject/MyProject.Web/MyProject.Web.csproj""
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: PowerShell test task
-  run: MyProject/BuildVersion.ps1 -ProjectFile ""MyProject/MyProject.Web/MyProject.Web.csproj""
-  shell: powershell
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void ScriptWithMultilineIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- script: |
-    echo Add other tasks to build, test, and deploy your project.
-    echo See https://aka.ms/yaml
-  displayName: 'Run a multi-line script'";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Run a multi-line script
-  run: |
-    echo Add other tasks to build, test, and deploy your project.
-    echo See https://aka.ms/yaml
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void PowershellWithConditionIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: PowerShell@2
-  displayName: 'PowerShell test task'
-  condition: and(eq('ABCDE', 'BCD'), ne(0, 1))
-  inputs:
-    targetType: inline
-    script: Write-Host 'Hello World'";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: PowerShell test task
-  run: Write-Host 'Hello World'
-  shell: powershell
-  if: (('ABCDE' == 'BCD') && (0 != 1))
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void AzureCLIIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: AzureCLI@2
-  displayName: 'Deploy infrastructure with ARM templates'
-  inputs:
-    azureSubscription: 'Service Connection to Azure Portal'
-    scriptType: ps
-    scriptPath: $(build.artifactstagingdirectory)\drop\EnvironmentARMTemplate\PowerShell\DeployInfrastructureCore.ps1
-    arguments: -ResourceGroupName ""$(ResourceGroupName)"" 
-  enabled: true
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Deploy infrastructure with ARM templates
-  uses: azure/cli@v1.0.0
-  with:
-    inlineScript: ${{ github.workspace }}\drop\EnvironmentARMTemplate\PowerShell\DeployInfrastructureCore.ps1 -ResourceGroupName ""${{ env.ResourceGroupName }}""
-    azcliversion: latest
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void AzureKeyVaultIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: AzureKeyVault@1
-  displayName: 'Get Azure Keyvault secrets' 
-  inputs:
-    azureSubscription: 'Service Connection to Azure Portal'
-    keyVaultName: myKeyvaultName
-    secretsFilter: '*'
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Get Azure Keyvault secrets
-  uses: azure/cli@v1.0.0
-  with:
-    inlineScript: az keyvault secret list --subscription Service Connection to Azure Portal --vault-name myKeyvaultName
-    azcliversion: latest
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void AzurePowershellIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: AzurePowerShell@4 
-  displayName: 'Run Azure PowerShell' 
-  inputs:
-    azureSubscription: 'Service Connection to Azure Portal'
-    ScriptPath: '$(build.artifactstagingdirectory)/drop/EnvironmentARMTemplate/PowerShell/Cleanup.ps1'
-    ScriptArguments: '-ResourceGroupName ""$(ResourceGroupName)""'
-    azurePowerShellVersion: LatestVersion
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Run Azure PowerShell
-  uses: azure/powershell@v1
-  with:
-    inlineScript: ${{ github.workspace }}/drop/EnvironmentARMTemplate/PowerShell/Cleanup.ps1 -ResourceGroupName ""${{ env.ResourceGroupName }}""
-    azPSVersion: latest
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void AzurePowershellWithOtherVersionIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: AzurePowerShell@4 
-  displayName: 'Run Azure PowerShell' 
-  inputs:
-    azureSubscription: 'Service Connection to Azure Portal'
-    ScriptPath: '$(build.artifactstagingdirectory)/drop/EnvironmentARMTemplate/PowerShell/Cleanup.ps1'
-    ScriptArguments: '-ResourceGroupName ""$(ResourceGroupName)""'
-    azurePowerShellVersion: OtherVersion
-    preferredAzurePowerShellVersion: 1.2.3
-    errorActionPreference: stop
-    FailOnStandardError: false
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Run Azure PowerShell
-  uses: azure/powershell@v1
-  with:
-    inlineScript: ${{ github.workspace }}/drop/EnvironmentARMTemplate/PowerShell/Cleanup.ps1 -ResourceGroupName ""${{ env.ResourceGroupName }}""
-    azPSVersion: 1.2.3
-    errorActionPreference: stop
-    failOnStandardError: false
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void AzurePowershellWithOtherVersionAndTargetAzurePsIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: AzurePowerShell@4 
-  displayName: 'Run Azure PowerShell' 
-  inputs:
-    azureSubscription: 'Service Connection to Azure Portal'
-    ScriptPath: '$(build.artifactstagingdirectory)/drop/EnvironmentARMTemplate/PowerShell/Cleanup.ps1'
-    ScriptArguments: '-ResourceGroupName ""$(ResourceGroupName)""'
-    azurePowerShellVersion: OtherVersion
-    targetAzurePs: 1.2.3
-    errorActionPreference: stop
-    FailOnStandardError: false
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Run Azure PowerShell
-  uses: azure/powershell@v1
-  with:
-    inlineScript: ${{ github.workspace }}/drop/EnvironmentARMTemplate/PowerShell/Cleanup.ps1 -ResourceGroupName ""${{ env.ResourceGroupName }}""
-    azPSVersion: 1.2.3
-    errorActionPreference: stop
-    failOnStandardError: false
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void AzurePowershellWithOtherVersionAndCustomTargetAzurePsIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: AzurePowerShell@4 
-  displayName: 'Run Azure PowerShell' 
-  inputs:
-    azureSubscription: 'Service Connection to Azure Portal'
-    ScriptPath: '$(build.artifactstagingdirectory)/drop/EnvironmentARMTemplate/PowerShell/Cleanup.ps1'
-    ScriptArguments: '-ResourceGroupName ""$(ResourceGroupName)""'
-    azurePowerShellVersion: OtherVersion
-    customTargetAzurePs: 1.2.3
-    errorActionPreference: stop
-    FailOnStandardError: false
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Run Azure PowerShell
-  uses: azure/powershell@v1
-  with:
-    inlineScript: ${{ github.workspace }}/drop/EnvironmentARMTemplate/PowerShell/Cleanup.ps1 -ResourceGroupName ""${{ env.ResourceGroupName }}""
-    azPSVersion: 1.2.3
-    errorActionPreference: stop
-    failOnStandardError: false
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void AzureWebAppDeploymentIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: AzureRmWebAppDeployment@4
-  inputs:
-    connectionType: 'AzureRM'
-    azureSubscription: ""$(azureSubscription)""
-    appType: 'functionApp'
-    webAppName: ""$(functionappName)""
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- uses: Azure/webapps-deploy@v2
-  with:
-    app-name: ${{ env.functionappName }}
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void AzureWebAppIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: AzureWebApp@1
-  displayName: 'Deploy Azure Web App- app1'
-  inputs:
-    azureSubscription: $(azureServiceConnectionId)
-    appName: $(webAppName)
-    package: $(build.artifactstagingdirectory)/drop/MyProject.Web.zip
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Deploy Azure Web App- app1
-  uses: Azure/webapps-deploy@v2
-  with:
-    app-name: ${{ env.webAppName }}
-    package: ${{ github.workspace }}/drop/MyProject.Web.zip
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void PublishPipelineArtifactsIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: PublishPipelineArtifact@0
-  displayName: Store artifact
-  inputs:
-    artifactName: 'drop'
-    targetPath: 'MyProject/bin/release/netcoreapp2.2/publish/'
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Store artifact
-  uses: actions/upload-artifact@v2
-  with:
-    path: MyProject/bin/release/netcoreapp2.2/publish/
-    name: drop
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void DeployWebAppIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-    - task: AzureRmWebAppDeployment@3
-      displayName: 'Azure App Service Deploy: web site'
-      inputs:
-        azureSubscription: 'MyProject connection to Azure Portal'
-        WebAppName: $(WebsiteName)
-        DeployToSlotFlag: true
-        ResourceGroupName: $(ResourceGroupName)
-        SlotName: 'staging'
-        Package: '$(build.artifactstagingdirectory)/drop/MyProject.Web.zip'
-        TakeAppOfflineFlag: true
-        JSONFiles: '**/appsettings.json'  
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: 'Azure App Service Deploy: web site'
-  uses: Azure/webapps-deploy@v2
-  with:
-    app-name: ${{ env.WebsiteName }}
-    package: ${{ github.workspace }}/drop/MyProject.Web.zip
-    slot-name: staging
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
         [TestMethod]
         public void DownloadIndividualStepTest()
         {
@@ -951,286 +307,6 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
   with:
     name: WebApp
     path: ${{ github.workspace }}/bin
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void DeployAzureWebAppContainerIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-  - task: AzureWebAppContainer@1
-    displayName: 'Azure Web App on Container Deploy'
-    inputs:
-      azureSubscription: '$(AzureServiceConnectionId)'
-      appName: $(AppName)
-      imageName: '$(ACRFullName)/$(ACRImageName)'
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Azure Web App on Container Deploy
-  uses: Azure/webapps-deploy@v2
-  with:
-    app-name: ${{ env.AppName }}
-    images: ${{ env.ACRFullName }}/${{ env.ACRImageName }}
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void DeployFunctionAppContainerIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-  - task: AzureFunctionAppContainer@1
-    displayName: 'Azure Function App on Container Deploy'
-    inputs:
-      azureSubscription: '$(AzureServiceConnectionId)'
-      appName: $(AppName)
-      imageName: '$(ACRFullName)/$(ACRImageName)'
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Azure Function App on Container Deploy
-  uses: Azure/webapps-deploy@v2
-  with:
-    app-name: ${{ env.AppName }}
-    images: ${{ env.ACRFullName }}/${{ env.ACRImageName }}
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void DotNetCoreCLIRestoreIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: DotNetCoreCLI@2
-  displayName: Restore
-  inputs:
-    command: restore
-    projects: MyProject/MyProject.Models/MyProject.Models.csproj
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Restore
-  run: dotnet restore MyProject/MyProject.Models/MyProject.Models.csproj
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-
-        [TestMethod]
-        public void DotNetCoreCLIPushIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: DotNetCoreCLI@2
-  displayName: 'Push to GitHub Packages'
-  condition: and(succeeded(), startsWith(variables['build.sourceBranch'], 'refs/tags/')) # Only run this step when the trigger event is a tag push
-  inputs:
-    command: 'push'
-    packagesToPush: '$(Build.ArtifactStagingDirectory)/*.nupkg'
-    nuGetFeedType: 'external'
-    publishFeedCredentials: 'GitHub Packages'
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Push to GitHub Packages
-  run: dotnet nuget push ${{ github.workspace }}/*.nupkg --source ""github""
-  if: (success() && startsWith(github.ref, 'refs/tags/'))
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void DotNetCoreCLIBuildIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: DotNetCoreCLI@2
-  displayName: Build
-  inputs:
-    projects: MyProject/MyProject.Models/MyProject.Models.csproj
-    arguments: '--configuration $(BuildConfiguration)'
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Build
-  run: dotnet MyProject/MyProject.Models/MyProject.Models.csproj --configuration ${{ env.BuildConfiguration }}
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-
-
-        [TestMethod]
-        public void DotNetCoreCLIPublishIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: DotNetCoreCLI@2
-  displayName: Publish
-  inputs:
-    command: publish
-    publishWebProjects: false
-    projects: MyProject/MyProject.Models/MyProject.Models.csproj
-    arguments: '--configuration $(BuildConfiguration) --output $(build.artifactstagingdirectory)'
-    zipAfterPublish: false
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Publish
-  run: dotnet publish MyProject/MyProject.Models/MyProject.Models.csproj --configuration ${{ env.BuildConfiguration }} --output ${{ github.workspace }}
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void DotNetCoreCLIPublishMultiLineIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: DotNetCoreCLI@2
-  displayName: 'Publish dotnet projects'
-  inputs:
-    command: publish
-    publishWebProjects: false
-    projects: |
-      src/Project.Service/Project.Service.csproj
-      src/Project.Web/Project.Web.csproj
-    arguments: '--configuration Release --output $(build.artifactstagingdirectory) --runtime win-x64'
-    zipAfterPublish: true
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Publish dotnet projects
-  run: |
-    dotnet publish src/Project.Service/Project.Service.csproj --configuration Release --output ${{ github.workspace }} --runtime win-x64 
-    dotnet publish src/Project.Web/Project.Web.csproj --configuration Release --output ${{ github.workspace }} --runtime win-x64
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-
-
-        [TestMethod]
-        public void DotNetCoreCLIPackIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: DotNetCoreCLI@2
-  displayName: 'dotnet pack'
-  inputs:
-    command: pack
-    packagesToPack: MyProject/MyProject.Models/MyProject.Models.csproj
-    versioningScheme: byEnvVar
-    versionEnvVar: BuildVersion
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: dotnet pack
-  run: dotnet pack MyProject/MyProject.Models/MyProject.Models.csproj
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void DotNetCoreCLIInvalidIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: DotNetCoreCLI@2
-  displayName: 'dotnet build but no inputs'
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- # This DotNetCoreCLI task is misconfigured, inputs are required
-  name: dotnet build but no inputs
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void SwapSlotsIndividualStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-    - task: AzureAppServiceManage@0
-      displayName: 'Swap Slots: website'
-      inputs:
-        azureSubscription: 'MyProject connection to Azure Portal'
-        WebAppName: $(WebsiteName)
-        ResourceGroupName: $(ResourceGroupName)
-        SourceSlot: 'staging'      
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: 'Swap Slots: website'
-  uses: Azure/cli@v1.0.0
-  with:
-    inlineScript: az webapp deployment slot swap --resource-group ${{ env.ResourceGroupName }} --name ${{ env.WebsiteName }} --slot staging --target-slot production
 ";
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
@@ -1494,161 +570,27 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
         //}
 
         [TestMethod]
-        public void SQLAzureDacPacDeployStepTest()
+        public void InvalidStepIndividualStepTest()
         {
             //Arrange
             Conversion conversion = new Conversion();
-            string yaml = @"
-          - task: SqlAzureDacpacDeployment@1
-            displayName: 'Azure SQL dacpac publish'
-            inputs:
-              azureSubscription: 'my connection to Azure Portal'
-              ServerName: '$(databaseServerName).database.windows.net'
-              DatabaseName: '$(databaseName)'
-              SqlUsername: '$(databaseLoginName)'
-              SqlPassword: '$(databaseLoginPassword)'
-              DacpacFile: '$(build.artifactstagingdirectory)/drop/MyDatabase.dacpac'
-              additionalArguments: '/p:BlockOnPossibleDataLoss=true'   
-";
+            string yaml = "- task: invalid fake task";
 
             //Act
             ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
 
             //Assert
             string expected = @"
-- # ""Note: Connection string needs to be specified - this is different than Pipelines where the server, database, user, and password were specified separately. It's recommended you use secrets for the connection string.""
-  name: Azure SQL dacpac publish
-  uses: azure/sql-action@v1
-  with:
-    server-name: ${{ env.databaseServerName }}.database.windows.net
-    connection-string: ${{ secrets.AZURE_SQL_CONNECTION_STRING }}
-    dacpac-package: ${{ github.workspace }}/drop/MyDatabase.dacpac
-    arguments: /p:BlockOnPossibleDataLoss=true";
+- # ""Error: the step 'invalid fake task' does not have a conversion path yet""
+  run: |
+    echo ""Error: the step 'invalid fake task' does not have a conversion path yet""
+    #task: invalid fake task
+";
+
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
         }
-
-        [TestMethod]
-        public void AntStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: Ant@1
-  inputs:
-    workingDirectory: ''
-    buildFile: 'build.xml'
-    javaHomeOption: 'JDKVersion'
-    jdkVersionOption: '1.8'
-    jdkArchitectureOption: 'x64'
-    publishJUnitResults: true
-    testResultsFiles: '**/TEST-*.xml'  
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- run: ant -noinput -buildfile build.xml
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-
-        [TestMethod]
-        public void ArchiveFilesStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: ArchiveFiles@2
-  inputs:
-    rootFolderOrFile: '$(build.sourcesDirectory)'
-    includeRootFolder: false";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- # 'Note: This is a third party action and currently only supports Linux: https://github.com/marketplace/actions/create-zip-file'
-  uses: montudor/action-zip@v0.1.0
-  with:
-    args: zip -qq -r  ${{ github.workspace }}
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void ArmTemplateDeploymentStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-  - task: AzureResourceGroupDeployment@2
-    displayName: 'Deploy ARM Template to resource group'
-    inputs:
-      azureSubscription: 'connection to Azure Portal'
-      resourceGroupName: $(ResourceGroupName)
-      location: '[resourceGroup().location]'
-      csmFile: '$(build.artifactstagingdirectory)/drop/ARMTemplates/azuredeploy.json'
-      csmParametersFile: '$(build.artifactstagingdirectory)/drop/ARMTemplates/azuredeploy.parameters.json'
-      overrideParameters: '-environment $(AppSettings.Environment) -locationShort $(ArmTemplateResourceGroupLocation)' 
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- name: Deploy ARM Template to resource group
-  uses: Azure/cli@v1.0.0
-  with:
-    inlineScript: az deployment group create --resource-group ${{ env.ResourceGroupName }} --template-file ${{ github.workspace }}/drop/ARMTemplates/azuredeploy.json --parameters  ${{ github.workspace }}/drop/ARMTemplates/azuredeploy.parameters.json -environment ${{ env.AppSettings.Environment }} -locationShort ${{ env.ArmTemplateResourceGroupLocation }}
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void ArmTemplateV3DeploymentStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: AzureResourceManagerTemplateDeployment@3
-  inputs:
-    deploymentScope: 'Resource Group'
-    azureResourceManagerConnection: 'copy-connection'
-    subscriptionId: '00000000-0000-0000-0000-000000000000'
-    action: 'Create Or Update Resource Group'
-    resourceGroupName: $(ResourceGroupName)
-    location: 'West US'
-    templateLocation: 'URL of the file'
-    csmFileLink: '$(AzureFileCopy.StorageContainerUri)templates/mainTemplate.json$(AzureFileCopy.StorageContainerSasToken)'
-    csmParametersFileLink: '$(AzureFileCopy.StorageContainerUri)templates/mainTemplate.parameters.json$(AzureFileCopy.StorageContainerSasToken)'
-    deploymentMode: 'Incremental'
-    deploymentName: 'deploy1'
-    overrideParameters: '-environment $(AppSettings.Environment) -locationShort $(ArmTemplateResourceGroupLocation)' 
-";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- uses: Azure/cli@v1.0.0
-  with:
-    inlineScript: az deployment group create --resource-group ${{ env.ResourceGroupName }} --template-file ${{ env.AzureFileCopy.StorageContainerUri }}templates/mainTemplate.json${{ env.AzureFileCopy.StorageContainerSasToken }} --parameters  -environment ${{ env.AppSettings.Environment }} -locationShort ${{ env.ArmTemplateResourceGroupLocation }}
-";
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-
+     
         //        [TestMethod]
         //        public void MSBuildStepTest()
         //        {
@@ -1776,58 +718,6 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
         }
 
         [TestMethod]
-        public void MSBuildStepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-        - task: VSBuild@1
-          inputs:
-            solution: '$(solution)'
-            msbuildArgs: '/p:DeployOnBuild=true /p:WebPublishMethod=Package /p:PackageAsSingleFile=true /p:SkipInvalidConfigurations=true /p:DesktopBuildPackageLocation=""$(build.artifactStagingDirectory)\WebApp.zip"" /p:DeployIisAppPath=""Default Web Site""'
-            platform: '$(buildPlatform)'
-            configuration: '$(buildConfiguration)'
-        ";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- run: msbuild '${{ env.solution }}' /p:configuration='${{ env.buildConfiguration }}' /p:platform='${{ env.buildPlatform }}' /p:DeployOnBuild=true /p:WebPublishMethod=Package /p:PackageAsSingleFile=true /p:SkipInvalidConfigurations=true /p:DesktopBuildPackageLocation=""${{ github.workspace }}\WebApp.zip"" /p:DeployIisAppPath=""Default Web Site""";
-
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
-        public void MSBuild2StepTest()
-        {
-            //Arrange
-            Conversion conversion = new Conversion();
-            string yaml = @"
-- task: MSBuild@1
-  inputs:
-    solution: '**/*.sln'
-    msbuildArchitecture: 'x86'
-    platform: 'Any CPU'
-    configuration: 'Release'
-    msbuildArguments: '/t:Publish /p:PublishUrl=""publish""'
-        ";
-
-            //Act
-            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
-
-            //Assert
-            string expected = @"
-- run: msbuild '**/*.sln' /p:configuration='Release' /p:platform='Any CPU' /t:Publish /p:PublishUrl=""publish""
-";
-
-            expected = UtilityTests.TrimNewLines(expected);
-            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }
-
-        [TestMethod]
         public void PublishTestResultsStepTest()
         {
             //Arrange
@@ -1943,8 +833,6 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
         //            expected = UtilityTests.TrimNewLines(expected);
         //            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
         //        }    
-
-
 
     }
 }
